refactor: start moving static functions into separate file
