+++ test/testPublishCmd.js
@@ -77,6 +77,13 @@ describe('Dynamic Strain (VCL) generation', () => {
     // console.log(PublishCommand.getStrainResolutionVCL(strainfile));
     assert.equal(vclfile.trim(), PublishCommand.getStrainResolutionVCL(strainfile).trim());
   });
+
+  it('getStrainResolutionVCL generates VLC for proxy strains', () => {
+    const strainfile = strainconfig.load(fs.readFileSync(path.resolve(__dirname, 'fixtures/proxystrains.yaml'), 'utf-8'));
+    const vclfile = fs.readFileSync(path.resolve(__dirname, 'fixtures/proxystrains.vcl')).toString();
+    // console.log(PublishCommand.getStrainResolutionVCL(strainfile));
+    assert.equal(vclfile.trim(), PublishCommand.getStrainResolutionVCL(strainfile).trim());
+  });
 });
 
 describe('hlx publish (Integration)', function suite() {
